Apply review comment by @AnastasiiaSergienko
